% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girt.R
\name{dgbeta}
\alias{dgbeta}
\title{Generalized Beta Distribution Beta Distribution in min < x < max}
\usage{
dgbeta(x, paramab, rangex)
}
\arguments{
\item{x}{A vector consisting of the random variable.}

\item{paramab}{A vector consisting of a and b parameters.}

\item{rangex}{A vector consisting of min and max of the random variable.}
}
\description{
Generalized Beta Distribution Beta Distribution in min < x < max
}
\author{
Shin-ichi Mayekawa \href{mailto:mayekawa@nifty.com}{mayekawa@nifty.com}
}
