% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girt.R
\name{dinvchi}
\alias{dinvchi}
\title{The density of chi inv dist
inverse chi distribution}
\usage{
dinvchi(phi, v = 1, tau = 1)
}
\arguments{
\item{phi}{phi. upper 0.}

\item{v}{a parameter.}

\item{tau}{a parameter.}
}
\description{
The density of chi inv dist
inverse chi distribution
}
