% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estdist}
\alias{estdist}
\title{Estimation of population distribution via EM algorithm.}
\usage{
estdist(x, para, N = 31L, eMLL = 1e-06, eDIST = 1e-04, fc0 = 2L,
  ng = 1L, gc0 = 2L, D = 1.702, fix = 1L, print = 0L, max = 4,
  min = -4, maxiter_em = 200L, rm_list = as.character(c("NONE")))
}
\arguments{
\item{x}{Item response matrix.}

\item{para}{item parameter data.frame estimated by \code{\link{estip}}}

\item{N}{the number of nodes in integration.}

\item{eMLL}{a convergence criteria(CC) for marginal log likelihood.}

\item{eDIST}{a CC for population distribution.}

\item{fc0}{a column of first item response.}

\item{ng}{the number of groups}

\item{gc0}{a column of group. the element must be integer and the minimum number must be 1.}

\item{D}{factor constant.}

\item{fix}{Don't use. If 1, fix population distribution mean and sigma each EM cycle.}

\item{print}{How much information you want to display? from 1 to 3. The larger, more information is displayed.}

\item{max}{maximum value of theta in integration.}

\item{min}{minimum value of theta in integration.}

\item{maxiter_em}{maximum iteration time for EM cycle.}

\item{rm_list}{a vector of item U want to remove for estimation. NOT list.}
}
\description{
This function is new version of \code{MML_EM_dist}
}
