% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Estep_girt}
\alias{Estep_girt}
\title{Estep in girt.}
\usage{
Estep_girt(x, a0, b0, Xq, AX, Yr, BY, D)
}
\arguments{
\item{x}{DataFrame.}

\item{a0}{discrinimation parameter vector.}

\item{b0}{difficulty parameter vector.}

\item{Xq}{node of theta dist.}

\item{AX}{weight of theta dist.}

\item{Yr}{node of phi dist.}

\item{BY}{weighit of phi dist.}

\item{D}{factor constant.}
}
\description{
Estep in girt.
}
