% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Estep_girt_mg}
\alias{Estep_girt_mg}
\title{Estep in girt.}
\usage{
Estep_girt_mg(x, a0, b0, Xq, AX, Yr, BY, D, group, ind, resp, MLL)
}
\arguments{
\item{x}{DataFrame.}

\item{a0}{discrinimation parameter vector.}

\item{b0}{difficulty parameter vector.}

\item{Xq}{node of theta dist.}

\item{AX}{weight of theta dist.}

\item{Yr}{node of phi dist.}

\item{BY}{weighit of phi dist.}

\item{D}{factor constant.}

\item{group}{a vector.}

\item{ind}{a design matrix for group.}

\item{resp}{a design matrix for person.}

\item{MLL}{a vector}
}
\description{
Estep in girt.
}
