% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Estep_irt}
\alias{Estep_irt}
\title{Estep in girt.}
\usage{
Estep_irt(xall, t0, Xm, Wm, group, ind, resp, D, MLL)
}
\arguments{
\item{xall}{Item response matrix.}

\item{t0}{item parameter vector}

\item{Xm}{node of theta dist.}

\item{Wm}{weight of theta dist.}

\item{group}{a vector.}

\item{ind}{a design matrix for group.}

\item{resp}{a design matrix for person.}

\item{D}{factor constant.}

\item{MLL}{a vector}
}
\description{
Estep in girt.
}
