% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girt.R
\name{estGip}
\alias{estGip}
\title{Genelar Item Response Theory parameter estimation}
\usage{
estGip(x, fc = 3, IDc = 1, Ntheta = 31, Nphi = 31,
  engine = "Cpp", eEM = 0.001, eMLL = 0.001, maxiter_em = 100,
  method = "L-BFGS-B", phi_dist = "invchi", v = 1, tau = 1,
  mu_ph = 0, sigma_ph = 1, mu_th = 0, sigma_th = 1)
}
\arguments{
\item{x}{DataFrame.}

\item{fc}{the first column.}

\item{IDc}{the ID column.}

\item{Ntheta}{the number of the nodes of theta dist.}

\item{Nphi}{the number of the nodes of phi dist.}

\item{engine}{Estep calculation engine.\code{Cpp} is very faster than \code{R}.}

\item{eEM}{a convergence criterion of item parameters in EM cycle.}

\item{eMLL}{a convergence criterion of marginal log likelihood in EM cycle.}

\item{maxiter_em}{the number of iteration of EM cycle.}
}
\description{
Genelar Item Response Theory parameter estimation
}
