% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irtfun2-package.R
\name{estheta}
\alias{estheta}
\title{A estimation theta function.}
\usage{
estheta(xall, param, est = "EAP", nofrands = 10, method = "NR",
  file = "default", output = FALSE, IDc = 1, gc = 2, fc = 3,
  gh = TRUE, N = 31, D = 1.702, maxtheta = 6, mintheta = -6,
  mu = 0, sigma = 1, counter = 1000)
}
\arguments{
\item{xall}{a item response data}

\item{param}{a item parameter file.If class is df, parameter column starts from second column and the order is a, b, c.}

\item{est}{estimation method option.EAP,MAP,MLE,PVs.}

\item{nofrands}{the number of PVs.}

\item{method}{iteration method option in MLE and MAP. NR is Newton Rapthon, SANN is Simulated Aannealing, Brent is optimization using optimise function}

\item{file}{a name of output csv file.}

\item{output}{logical. If TRUE write csv file. defauli is FALSE.}

\item{IDc}{colomn of ID, default is 1.Even If df has no ID columns this function works.}

\item{gc}{column of group or population numbers. If df has no grouo ID, set 0.}

\item{fc}{column of first item response, default is 3.}

\item{gh}{logical. If TRUE, is default, gaussHermite quadrature runs for EAP estimation.}

\item{N}{the number of nodes. default is 31.}

\item{D}{a factor constant. deafault is 1.702.}

\item{maxtheta}{the maximul value of theta in integration.}

\item{mintheta}{the minimum value of theta in integration.}

\item{mu}{a hyperparameter of prior distribution.}

\item{sigma}{same as above.}

\item{counter}{Display progress every counter value.}
}
\value{
a list has ID, rawscore, theta, se, person fit index Z3 and log likelihood.
}
\description{
A estimation theta function.
}
