% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estip2.R
\name{estip2}
\alias{estip2}
\title{Unidimensional Item Response Theory parameter estimation}
\usage{
estip2(x, fc = 3, IDc = 1, Gc = NULL, bg = 1, Ntheta = 31,
  D = 1.702, method = "Fisher_Scoring", model = "2PL",
  max_func = "N", mu_th = 0, sigma_th = 1, min_th = -4,
  max_th = 4, eEM = 0.001, eMLL = 0.001, maxiter_em = 100,
  fix_a = 1, mu_a = 0, sigma_a = 1, mu_b = 0, sigma_b = 1,
  mu_c = 0, sigma_c = 1, w_c = 1)
}
\arguments{
\item{x}{DataFrame.}

\item{fc}{the first column.}

\item{IDc}{the ID column.}

\item{Gc}{the grade column.}

\item{bg}{a mumber of base grade.}

\item{Ntheta}{the number of the nodes of theta dist.}

\item{D}{a scale constant.}

\item{method}{the method of optimiser.  Default is "Fisher_Scoring", but \code{\link[stats]{optim}} function also be able to use.}

\item{model}{a model vector}

\item{max_func}{a character of object function. "N" is MML-EM, "B" is marginal Bayes.}

\item{mu_th}{a hyper parameter of normal dist for theta}

\item{sigma_th}{a hyper parameter of normal dist for theta}

\item{min_th}{a minimum value of theta distribution}

\item{max_th}{a maximum value of theta distribution}

\item{eEM}{a convergence criterion of item parameters in EM cycle.}

\item{eMLL}{a convergence criterion of marginal log likelihood in EM cycle.}

\item{maxiter_em}{the number of iteration of EM cycle.}

\item{fix_a}{a fix parameter for slope parameter of 1PLM}

\item{mu_a}{a hyper parameter for slope parameter prior distribution(lognormal) in marginal Bayes estimation.}

\item{sigma_a}{a hyper parameter for slope parameter prior distribution(lognormal) in marginal Bayes estimation.}

\item{mu_b}{a hyper parameter for location parameter prior distribution(normal) in marginal Bayes estimation.}

\item{sigma_b}{a hyper parameter for location parameter prior distribution(normal) in marginal Bayes estimation.}

\item{mu_c}{a hyper parameter for lower asymptote parameter prior distribution(beta) in marginal Bayes estimation.}

\item{sigma_c}{a hyper parameter for lower asymptote parameter prior distribution(beta) in marginal Bayes estimation.}

\item{w_c}{a weight parameter for lower asymptote parameter prior distribution(beta) in marginal Bayes estimation.}
}
\value{
the output is a list that has item parameters data.frame and these Standard Error.
}
\description{
Unidimensional Item Response Theory parameter estimation
}
\examples{
res1 <- estip2(x=sim_data_2, Gc=NULL, fc=2, Ntheta=21)
# check the parameters
res1$para
res1$SE

res2 <- estip2(x=sim_data_4, Gc=2, bg=2, fc=3, Ntheta=21)
}
