% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simgen.R
\name{Flol}
\alias{Flol}
\title{Calculate subject log likelihood.}
\usage{
Flol(theta, x, a, b, c)
}
\arguments{
\item{theta}{theta vector.}

\item{x}{item response matrix or df.}

\item{a}{slope parameter.}

\item{b}{location parameter.}

\item{c}{asymptote parameter.}
}
\description{
Calculate subject log likelihood.
}
