% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girt.R
\name{gptheta}
\alias{gptheta}
\title{ICC of GIRT model}
\usage{
gptheta(theta, phi, a, b, D)
}
\arguments{
\item{theta}{the location parameter of person}

\item{phi}{the disprecision parameter of person}

\item{a}{the discrimination parameter of item}

\item{b}{the difficulty parameter of item}

\item{D}{a scale constant}
}
\description{
ICC of GIRT model
}
