% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{ifind}
\alias{ifind}
\title{Item Fit indexes 'chi^2' and 'g^2' statistics}
\usage{
ifind(x, para, theta, fc = 3, H = 10, p = 0.05, D = 1.702)
}
\arguments{
\item{x}{data.frame of item response data.}

\item{para}{item parameter data.frame estimated by \code{\link{estip}}.}

\item{theta}{theta parameter vector. This length must be same to \code{x}.}

\item{fc}{a first column of item response data.frame.}

\item{H}{the munber of cut point. To know how to cut, see\code{\link[Hmisc]{cut2}}}

\item{p}{\emph{p} value.}

\item{D}{a scale constant}
}
\description{
Item Fit indexes 'chi^2' and 'g^2' statistics
}
\examples{
res <- ifind(sim_data_1, sim_param$para, sim_eap$res$EAP, fc=2)
# result
res$X2
res$G2
res$ggplot
}
