% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{ifind2}
\alias{ifind2}
\title{Item Fit indexes 'chi^2' and 'g^2' statistics that cut scores based on observed score.}
\usage{
ifind2(x, para, theta, Gc = 2, fc = 3, p = 0.05, D = 1.702)
}
\arguments{
\item{x}{data.frame of item response data.}

\item{para}{item parameter data.frame estimated by \code{\link{estip}}.}

\item{theta}{theta parameter vector. This length must be same to \code{x}.}

\item{Gc}{the munber of sub groups.}

\item{fc}{a first column of item response data.frame.}

\item{p}{\emph{p} value.}

\item{D}{a scale constant}
}
\description{
Item Fit indexes 'chi^2' and 'g^2' statistics that cut scores based on observed score.
}
