% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{ifind3}
\alias{ifind3}
\title{Item Fit indexes 'In Fit' and 'Out Fit' statistics, which is based on expexted and observed probability.}
\usage{
ifind3(x, para, theta, fc = 3, D = 1.702)
}
\arguments{
\item{x}{data.frame of item response data.}

\item{para}{item parameter data.frame estimated by \code{\link{estip}}.}

\item{theta}{theta parameter vector. This length must be same to \code{x}.}

\item{fc}{a first column of item response data.frame.}

\item{D}{a scale constant}
}
\description{
Item Fit indexes 'In Fit' and 'Out Fit' statistics, which is based on expexted and observed probability.
}
\examples{
res <- ifind3(sim_data_1, sim_param$para, sim_eap$res$EAP, fc=2)
res
}
