% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irtfun2-package.R
\name{iif}
\alias{iif}
\title{Item Information Function for IRT 1~3PLM}
\usage{
iif(theta, a, b, c, D)
}
\arguments{
\item{theta}{the person ability parameter}

\item{a}{the slope parameter}

\item{b}{the location parameter}

\item{c}{the guessing parameter}

\item{D}{a scale constant}
}
\description{
Item Information Function for IRT 1~3PLM
}
