% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girt.R
\name{invchi}
\alias{invchi}
\title{inverse chi distribution}
\usage{
invchi(phi, v = 3, tau = 1)
}
\arguments{
\item{phi}{phi. upper 0.}

\item{v}{a parameter.}

\item{tau}{a parameter.}
}
\description{
inverse chi distribution
}
