% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-tse.R
\name{irt_ytx}
\alias{irt_ytx}
\title{IRT true score equating
Reference: Kolen & Brennan (2014, p.194)}
\usage{
irt_ytx(paraX, paraY, D = 1.702)
}
\arguments{
\item{paraX}{df of itemparameter on Form X.}

\item{paraY}{df of itemparameter on Form Y.}

\item{D}{factor constant.}
}
\description{
IRT true score equating
Reference: Kolen & Brennan (2014, p.194)
}
