% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irtfun2-package.R
\docType{package}
\name{irtfun2-package}
\alias{irtfun2}
\alias{irtfun2-package}
\title{irtfun2: Some useful functions for IRT}
\description{
This package provedes functions for IRT analysis, for instance, estimation item and subject parameter. Need C++ compiler(Rtools or Xcode etc) to use a part of functions.
}
\author{
\strong{Maintainer}: Takumi Shibuya \email{sep10.taku.izum@gmail.com}

}
\keyword{internal}
