% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estheta.R
\name{LL}
\alias{LL}
\title{The log likelihood function of IRT 1~3PLM}
\usage{
LL(u, theta, a, b, c, D)
}
\arguments{
\item{u}{the item response pattern}

\item{theta}{the person ability parameter}

\item{a}{the slope parameter}

\item{b}{the location parameter}

\item{c}{the guessing parameter}

\item{D}{a scale constant}
}
\description{
The log likelihood function of IRT 1~3PLM
}
