% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estheta.R
\name{LL_b}
\alias{LL_b}
\title{The log likelihood function of Bayesian IRT 1~3PLM}
\usage{
LL_b(u, theta, a, b, c, mu, sigma, D)
}
\arguments{
\item{u}{the item response pattern}

\item{theta}{the person ability parameter}

\item{a}{the slope parameter}

\item{b}{the location parameter}

\item{c}{the guessing parameter}

\item{mu}{a hyperparameter of prior distribution(Gauss).}

\item{sigma}{same as above.}

\item{D}{a scale constant}
}
\description{
a prior distribution is Gaussian(normal) distribution
}
