% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estheta.R
\name{pfit}
\alias{pfit}
\title{Person fit index: z_3 statistics}
\usage{
pfit(dat, a, b, c, D)
}
\arguments{
\item{dat}{matrix which has theta estimates in the first column and the response data in the other.}

\item{a}{a vector item discrimination parameter.}

\item{b}{vector, item difficulty parameter.}

\item{c}{a vector. lower asymptote parameter.}

\item{D}{a scale constant.}
}
\value{
a matrix of z3 values
}
\description{
This function is a provisional version so will be update soon.
}
