% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mext-contract.R
\name{pfU}
\alias{pfU}
\title{Equipercentile equating function of row score.
This function generates the Form Y equipercentle equivalent of score x on FormX, eY(x).}
\usage{
pfU(p, tabley2)
}
\arguments{
\item{p}{percentile.}

\item{tabley2}{distribution table(must bedata.frame). \code{V2} is cumulative discrete distribution.}
}
\description{
Equipercentile equating function of row score.
This function generates the Form Y equipercentle equivalent of score x on FormX, eY(x).
}
\author{
Takumi Shibuya.
}
