% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irtfun2-package.R
\name{pitheta}
\alias{pitheta}
\title{MAP posterior information for IRT 1~3PLM}
\usage{
pitheta(dat, a, b, c, sigma, D)
}
\arguments{
\item{dat}{the matrix or data.frame which has theta and item response pattern vector}

\item{a}{the slope parameter}

\item{b}{the location parameter}

\item{c}{the guessing parameter}

\item{sigma}{a hyper parameter}

\item{D}{a scale constant}
}
\description{
MAP posterior information for IRT 1~3PLM
}
