% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mext-contract.R
\name{prf}
\alias{prf}
\title{percentile ranc function.}
\usage{
prf(q, table)
}
\arguments{
\item{q}{test raw score}

\item{table}{distribution table generated by \code{dist.f} function.}
}
\description{
percentile ranc function.
}
\author{
Takumi Shibuya.
}
