% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mext-contract.R
\name{probability}
\alias{probability}
\title{a function generates conditional probability for fixed ability on 2-parameter logistic model}
\usage{
probability(trait, a, b, c, D = 1.702)
}
\arguments{
\item{trait}{a vector of theta in IRT.}

\item{a}{a slope parameter}

\item{b}{a location parameter}

\item{c}{a lower asymptote parameter}

\item{D}{a factor constant}
}
\description{
a function generates conditional probability for fixed ability on 2-parameter logistic model
}
\references{
Kolen, M. J., & Brennan, R. L. (2014). Test Equating, Scaling, and Linking. Springer.
}
\author{
The original Fortran77 program was developed by Inoue,S., December 1990., extended by Shibayama,T., January 1991., translated into R by Shibayama,T. September 2008., functionalized by Itamiya, C., & Shibuya. T., June 2018.
}
