% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irtfun2-package.R
\name{ptheta}
\alias{ptheta}
\title{The ICC of IRT 1~3PLM}
\usage{
ptheta(theta, a, b, c, D = 1.702)
}
\arguments{
\item{theta}{the person ability parameter}

\item{a}{the slope parameter}

\item{b}{the location parameter}

\item{c}{the guessing parameter}

\item{D}{a scale constant}
}
\description{
The ICC of IRT 1~3PLM
}
