% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{Q3_stat}
\alias{Q3_stat}
\title{Item Fit indexes 'Q3'statistics. Diagnosis Local Item Independence(LID)}
\usage{
Q3_stat(
  x,
  para,
  theta,
  fc = 3,
  D = 1.702,
  use = "pairwise.complete.obs",
  abs_value = 0.2
)
}
\arguments{
\item{x}{data.frame of item response data.}

\item{para}{item parameter data.frame estimated by \code{\link{estip}}.}

\item{theta}{theta parameter vector. This length must be same to \code{x}.}

\item{fc}{a first column of item response data.frame.}

\item{D}{a scale constant}

\item{use}{an optional character string giving a method for computing covariances in the presence of missing values. See \code{\link{cor}}}

\item{abs_value}{an absolute value for Q3.}
}
\description{
Item Fit indexes 'Q3'statistics. Diagnosis Local Item Independence(LID)
}
