% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simgen.R
\name{resfunc}
\alias{resfunc}
\title{Generate binary data from response probability matrix.}
\usage{
resfunc(prob, power)
}
\arguments{
\item{prob}{response probabirity matrix}

\item{power}{a power of probability of NA. for example power = 1/5}
}
\description{
Generate binary data from response probability matrix.
}
