% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girt.R
\name{rinvchi}
\alias{rinvchi}
\title{The random number of chi inv dist}
\usage{
rinvchi(n, max = 5, v = 1, tau = 1)
}
\arguments{
\item{n}{the numeber of random variable to generate}

\item{v}{hyper parameter}

\item{tau}{hyper parameter}

\item{maz}{max ov vector}
}
\description{
The random number of chi inv dist
}
