% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set.R
\docType{data}
\name{sim_dat_girt}
\alias{sim_dat_girt}
\title{Simulated multi group item response data for GIRT model}
\format{See \code{\link{estGip}}.}
\usage{
sim_dat_girt
}
\description{
A list of estimated result from \code{\link{estGip}}.
}
\examples{
# GIRT estimated parameter
set.seed(0204)
theta <- rnorm(3000)
phi <- rinvchi(3000, max = 2)
a <- rlnorm(30, sdlog = 0.25)
b <- rnorm(30)
#sim_dat_girt <- sim_gen(theta=theta, phi=phi, a=a, b=b)
}
\keyword{datasets}
