% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set.R
\docType{data}
\name{sim_dat_st}
\alias{sim_dat_st}
\title{Simulated multi group item response data for scaling test design.}
\format{a dataset with 72 cols, first col is ID line and second col is group ID line.}
\usage{
sim_dat_st
}
\description{
This data set contains 70 item responses of 2,500 subjects for scaling test design.
}
\keyword{datasets}
