% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set.R
\docType{data}
\name{sim_data_4}
\alias{sim_data_4}
\title{Simulated multi group item response data for 2PLM}
\format{a dataset with 32 cols, first col is ID line and second col is group ID line.}
\usage{
sim_data_4
}
\description{
This data set contains 30 item responses of 9,000 subjects of 3 diffelent level groups.
}
\keyword{datasets}
