% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set.R
\docType{data}
\name{sim_eap}
\alias{sim_eap}
\title{Estimated theta parameter under 2PLM.}
\format{See \code{\link{estheta}}.}
\usage{
sim_eap
}
\description{
A list of estimated result from \code{\link{estheta}}.
}
\examples{
head(sim_eap$res)
}
\keyword{datasets}
