% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set.R
\docType{data}
\name{sim_param}
\alias{sim_param}
\title{Estimated 2PLM item parameter.}
\format{See \code{\link{estip}}.}
\usage{
sim_param
}
\description{
A list of estimated result from \code{\link{estip}}.
}
\examples{
sim_param$para
sim_param$SE
}
\keyword{datasets}
