% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irtfun2-package.R
\name{stat_pv}
\alias{stat_pv}
\title{Calculate the imputation variance}
\usage{
stat_pv(PVs_only)
}
\arguments{
\item{PVs_only}{a df given by the result of \code{estheta(est="PVs")}.}
}
\description{
Calculate the imputation variance
}
