% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simgen.R
\name{subfunc}
\alias{subfunc}
\title{Generate binary data from response probability vector}
\usage{
subfunc(prob, power)
}
\arguments{
\item{prob}{response probabirity vector}

\item{power}{a power of probability of NA. for example power = 1/5}
}
\description{
Generate binary data from response probability vector
}
