% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{theta_pv}
\alias{theta_pv}
\title{Sampling plausible values(PVs) based on rejection sampling.}
\usage{
theta_pv(
  x,
  nofrands,
  eap_apply,
  const_apply,
  map_apply,
  n,
  maxtheta,
  mintheta,
  a,
  b,
  c,
  D,
  mu,
  sigma
)
}
\arguments{
\item{x}{response data}

\item{nofrands}{the number of PVs}

\item{eap_apply}{a numeric vector of estimated EAP.}

\item{const_apply}{a numeric vector of const of a posterior distribution.}

\item{map_apply}{a numeric vector of estimated MAP.}

\item{n}{the number of subjects.}

\item{maxtheta}{a max value of theta.}

\item{mintheta}{a mininum value of theta.}

\item{a}{an item slope parameter vector.}

\item{b}{an item location parameter vector.}

\item{c}{an item asymptote parameter vector.}

\item{D}{a factor constant.}

\item{mu}{p hyper parameter of a posterior distribution.}

\item{sigma}{same as above.}
}
\description{
Sampling plausible values(PVs) based on rejection sampling.
}
