% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irtfun2-package.R
\name{tif}
\alias{tif}
\title{Test Information Function for IRT 1~3PLM}
\usage{
tif(theta, a, b, c, D)
}
\arguments{
\item{theta}{the person ability parameter}

\item{a}{the slope parameter}

\item{b}{the location parameter}

\item{c}{the guessing parameter}

\item{D}{a scale constant}
}
\description{
Test Information Function for IRT 1~3PLM
}
