% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omake.R
\name{timer}
\alias{timer}
\title{Measure the time and display original message.}
\usage{
timer(
  StopTime,
  MESSAGE = "STOP!!",
  disp = T,
  units = "secs",
  cex = 5,
  font = 1,
  col = 1,
  family = ""
)
}
\arguments{
\item{StopTime}{How long time you want to measure?}

\item{MESSAGE}{What message U want to display when stop timer.}

\item{disp}{logical. If FALSE, don't display the message.}

\item{units}{time units. auto", "secs", "mins", "hours", "days", "weeks". The default is "secs".}

\item{cex}{graphics parameter. expansion rate.}

\item{font}{graphics parameter. font number.}

\item{col}{graphics parameter. font colour.}

\item{family}{graohics parameter.}
}
\description{
Measure the time and display original message.
}
