% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set.R
\docType{data}
\name{true_param}
\alias{true_param}
\title{true 2PLM item parameter.}
\format{See \code{\link{sim_param}}.}
\usage{
true_param
}
\description{
A list of true parameter .
}
\examples{
set.seed(0204)
theta <- rnorm(3000)
a <- rlnorm(30, sdlog = 0.25)
b <- rnorm(30)
#list(theta,a,b)
}
\keyword{datasets}
