% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set.R
\docType{data}
\name{true_param_st}
\alias{true_param_st}
\title{true 2PLM item parameter of 'sim_dat_st'.}
\format{See \code{\link{sim_dat_st}}.}
\usage{
true_param_st
}
\description{
A list of true parameter .
}
\keyword{datasets}
