% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mext-contract.R
\name{tscore_dist}
\alias{tscore_dist}
\title{A function calculates IRT true score.}
\usage{
tscore_dist(theta, a, b, c, D = 1.702)
}
\arguments{
\item{theta}{a vector of theta estimator EAP, MAP or MLE...}

\item{a}{a slope parameter.}

\item{b}{a location parameter}

\item{c}{a lower asymptote parameter}

\item{D}{a factor constant}
}
\description{
A function calculates IRT true score.
}
