% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-tse.R
\name{tse}
\alias{tse}
\title{Newton-Raphton method in IRT true score equating
Reference: Kolen & Brennan (2014, p.194)}
\usage{
tse(tau, t0, a, b, c, D)
}
\arguments{
\item{tau}{true score, integer, vector.}

\item{t0}{initial value of theta in NR.}

\item{a}{slope parameter vector.}

\item{b}{location parameter vector.}

\item{c}{asymptote parameter vector.}

\item{D}{factor constant.}
}
\description{
Newton-Raphton method in IRT true score equating
Reference: Kolen & Brennan (2014, p.194)
}
