% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irtfun2-package.R
\name{CEquating}
\alias{CEquating}
\title{Equating item parameter only sampled by common item design. FOR ONLY 2PLM}
\usage{
CEquating(T_para, F_para, method = "SL", D = 1.702, N = 31,
  mintheta = -6, maxtheta = 6, output = F, Fname = "default",
  Change = 0, Easy = F)
}
\arguments{
\item{T_para}{a parameter file equating TO.}

\item{F_para}{a parameter file equating FROM.}

\item{method}{equating method. U can select "SL" or "HB".}

\item{D}{a factor canstant.}

\item{N}{the number of nodes.}

\item{mintheta}{a minimum value of theta in integration.}

\item{maxtheta}{a maximum value of theta in integration.}

\item{output}{logical. if TRUE output CSV file.}

\item{Fname}{character. a file name of csv file.}

\item{Change}{int. if 0 equated parameter of common item is no transformed parameter from T_para, if 2 transformed parameter, if1 mean and geometric mean.}

\item{Easy}{logical. if TRUE output file is adaptive for Easy Estimation.}
}
\description{
Equating item parameter only sampled by common item design. FOR ONLY 2PLM
}
