% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mext-contract.R
\name{dist.f}
\alias{dist.f}
\title{a table for Equipercentile equating function of row score.
The return table contains roe score, frequency, cumrative frequency, percent, cumlative percent, discrete distribution and cumlative discrete distribution.}
\usage{
dist.f(x)
}
\arguments{
\item{x}{test score vector}
}
\description{
a table for Equipercentile equating function of row score.
The return table contains roe score, frequency, cumrative frequency, percent, cumlative percent, discrete distribution and cumlative discrete distribution.
}
\author{
Takumi Shibuya.
}
