% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mext-contract.R
\name{epe}
\alias{epe}
\title{Equipercentile equating function of row score.
This function generates the Form Y equipercentle equivalent of score x on FormX, eY(x).}
\usage{
epe(x, y)
}
\arguments{
\item{x}{test score of Form X}

\item{y}{test score of Form Y}
}
\description{
Equipercentile equating function of row score.
This function generates the Form Y equipercentle equivalent of score x on FormX, eY(x).
}
