% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girt.R
\name{estGip}
\alias{estGip}
\title{Generalized Item Response Theory parameter estimation}
\usage{
estGip(x, fc = 3, Gc = NULL, bg = 1, IDc = 1, D = 1.702,
  Ntheta = 31, Nphi = 10, method = "BFGS", th_dist = "normal",
  rm_list = NULL, phi_dist = "uniform", v = 3, tau = 1,
  mu_ph = 0, sigma_ph = 0.25, min_ph = 0.01, max_ph = 2,
  paramab = c(1, 4), mu_th = 0, sigma_th = 1, min_th = -4,
  max_th = 4, eEM = 0.001, eMLL = 1e-06, eDIST = 1e-04,
  maxiter_em = 100, print = 0, esteap = FALSE, estdist = FALSE,
  initial = NULL)
}
\arguments{
\item{x}{DataFrame.}

\item{fc}{the first column.}

\item{Gc}{the grade column.}

\item{bg}{a mumber of base grade.}

\item{IDc}{the ID column.}

\item{D}{scale constant}

\item{Ntheta}{the number of the nodes of theta dist.}

\item{Nphi}{the number of the nodes of phi dist.}

\item{method}{the method of optimiser.  Default is "Fisher_Scoring", but \code{\link[stats]{optim}} function also be able to use.}

\item{th_dist}{a type of theta dist."normal" or "empirical"}

\item{rm_list}{a vector of item U want to remove for estimation. NOT list.}

\item{phi_dist}{a prior distribution of phi. \code{invchi} is inverse chi distribution. \code{lognormal} is log normal distribution.}

\item{v}{a hyper parameter of invchi for phi}

\item{tau}{a hyper parameter of invchi for phi}

\item{mu_ph}{a hyper parameter of lognormal dist for phi}

\item{sigma_ph}{a hyper parameter of lognormal dist for phi}

\item{min_ph}{a minimum value of phi distribution}

\item{max_ph}{a maximum value of phi distribution}

\item{paramab}{a hyper parameter of generalized beta distribution}

\item{mu_th}{a hyper parameter of normal dist for theta}

\item{sigma_th}{a hyper parameter of normal dist for theta}

\item{min_th}{a minimum value of theta distribution}

\item{max_th}{a maximum value of theta distribution}

\item{eEM}{a convergence criterion related to item parameters in EM cycle.}

\item{eMLL}{a convergence criterion related to negative twice log likelihood in EM cycle.}

\item{eDIST}{a convergence criterion related to the parameter of theta distribution in EM cycle.}

\item{maxiter_em}{the number of iteration of EM cycle.}

\item{print}{How much information you want to display? from 1 to 3. The larger, more information is displayed.}

\item{esteap}{logical. If \code{TRUE}, estimate subject theta & phi EAP.}

\item{estdist}{logical. If \code{TRUE}, estimate population distribution theta & phi.}

\item{initial}{if use specify initial value, set here data.frame.}
}
\value{
the output is a list that has item parameter and person parameter.
}
\description{
Generalized Item Response Theory parameter estimation
}
\examples{
res <- estGip(x=sim_dat_girt,fc=2, Ntheta=10, Nphi = 5, min_ph = 0.001, max_ph = 2)
# check the parameters
res$item
head(res$person)

}
