% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estip}
\alias{estip}
\title{Estimate item parameter for binary{0,1} response data.
1PL,2PL,3PL,Bayes1PL,Bayes2PL and multigroup estimation is avairable now. U must install C++ compiler(Rtools for windows or Xcode for Mac)in your PC or Mac.}
\usage{
estip(x, model = "2PL", N = 31L, bg0 = 1L, eMLL = 1e-06,
  eEM = 1e-04, eM = 0.001, emu = 0.001, esd = 0.01, fc0 = 2L,
  ng = 1L, gc0 = 2L, D = 1.702, fix = 1L, print = 0L, ic = 1/5,
  max = 6, min = -6, mu = 0, sigma = 1, Bayes = 0L, mu_a = 0,
  sigma_a = 1, mu_b = 0, sigma_b = 2, mu_c = 4/13, w_c = 13,
  min_a = 0.1, maxabs_b = 20, maxiter_em = 200L, maxiter_j = 20L,
  maxskip_j = 5L, rm_list = as.character(c("NONE")),
  thdist = "normal", e_ell = 1L, EM_dist = 1L)
}
\arguments{
\item{x}{an item response data which class is data.frame object.}

\item{model}{Character.U can select which one, "1PL","2PL","3PL".}

\item{N}{the number of nodes in integration.}

\item{bg0}{the number of base grade.}

\item{eMLL}{a convergence criteria(CC) forf marginal log likelihood.}

\item{eEM}{a CC in EM cycle.}

\item{eM}{a CC in M step.}

\item{emu}{a CC for population distribution mean.}

\item{esd}{a CC for population distribution standard deviation.}

\item{fc0}{a column of first item response.}

\item{ng}{the number of groups}

\item{gc0}{a column of group. the element must be integer and the minimum number must be 1.}

\item{D}{a scaling constant.}

\item{fix}{Don't use. If 1, fix population distribution mean and sigma each EM cycle.}

\item{print}{How much information you want to display? from 1 to 3. The larger, more information is displayed.}

\item{ic}{initial value for guessing parameter.}

\item{max}{maximum value of theta in integration.}

\item{min}{minimum value of theta in integration.}

\item{mu}{hyperparameter for theta dist.}

\item{sigma}{same as above.}

\item{Bayes}{If 1, marginal Bayesian estimation runs. This option is unstable. Don't use.}

\item{mu_a}{hyperparameter of log normal dist for slope parameter.}

\item{sigma_a}{same as above.}

\item{mu_b}{hyperparameter of normal dist for location parameter.}

\item{sigma_b}{same as above.}

\item{mu_c}{hyperparameter for lower asymptote parameter.}

\item{w_c}{weight of a Beta dist.}

\item{min_a}{minimum value of slope parameter.}

\item{maxabs_b}{maximum absolute value of location parameter.}

\item{maxiter_em}{maximum iteration time for EM cycle.}

\item{maxiter_j}{maximum iteration time for Newton Raphton in M step.}

\item{maxskip_j}{Dont use.}

\item{rm_list}{a vector of item U want to remove for estimation. NOT list.}

\item{thdist}{Which distribution do you want \code{normal} or \code{empirical} for E step.}

\item{e_ell}{If 1 use a CC of expected log likelihood.}

\item{EM_dist}{If 1, calculate esimated population distribution via EM argorithm.}
}
\description{
Estimate item parameter for binary{0,1} response data.
1PL,2PL,3PL,Bayes1PL,Bayes2PL and multigroup estimation is avairable now. U must install C++ compiler(Rtools for windows or Xcode for Mac)in your PC or Mac.
}
