% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mext-contract.R
\name{pfU}
\alias{pfU}
\title{The inverse of percentile rank function which uses the smallest integer score with cumulative persent that is greater then \code{p}.
This function generates \code{p} percentile of the score in \code{tabley2}.}
\usage{
pfU(p, tabley2)
}
\arguments{
\item{p}{percentile.}

\item{tabley2}{distribution table(must be data.frame). \code{V2} is cumulative discrete distribution.}
}
\description{
The inverse of percentile rank function which uses the smallest integer score with cumulative persent that is greater then \code{p}.
This function generates \code{p} percentile of the score in \code{tabley2}.
}
\author{
Takumi Shibuya.
}
