% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mext-contract.R
\name{prf}
\alias{prf}
\title{percentile rank function.}
\usage{
prf(q, table)
}
\arguments{
\item{q}{test raw score vector.}

\item{table}{This must be distribution table generated by \code{dist.f} function.}
}
\description{
percentile rank function.
}
\author{
Takumi Shibuya.
}
