% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set.R
\docType{data}
\name{sim_dat_girt}
\alias{sim_dat_girt}
\title{Simulated item response data for GIRT model}
\format{a dataset with 31 cols, first col is ID line.See \code{\link{estGip}}.}
\usage{
sim_dat_girt
}
\description{
This data set contains 30 item responses of 3,000 subjects for GIRT model
}
\examples{
# GIRT estimated parameter
set.seed(0204)
theta <- rnorm(3000)
phi <- rinvchi(3000, max = 2)
a <- rlnorm(30, sdlog = 0.25)
b <- rnorm(30)
sim_dat_girt <- sim_gen(theta=theta, phi=phi, a=a, b=b)
}
\keyword{datasets}
