% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set.R
\docType{data}
\name{sim_data_1}
\alias{sim_data_1}
\title{Simulated item response data for 1PLM}
\format{
a dataset with 31 cols, first col is ID line.
}
\usage{
sim_data_1
}
\description{
This data set contains 30 item responses of 3,000 subjects
}
\examples{
set.seed(0204)
# single test data: subjects=3,000 and item=30, 1PLM
sim_data_1 <- irtfun2::sim_gen(theta=rnorm(3000), b=rnorm(30), a=rep(1,30))
}
\keyword{datasets}
