% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simgen.R
\name{sim_gen}
\alias{sim_gen}
\title{Generate simulation binary data in IRT.}
\usage{
sim_gen(theta, a, b, c, item = "A", power = 0)
}
\arguments{
\item{theta}{theta vector}

\item{a}{slope parameter.}

\item{b}{location parameter.}

\item{c}{asymptote parameter.}

\item{item}{Character. item code.}

\item{power}{a power of probability of NA. for example power = 1/5}
}
\description{
This function contain \code{resfunc} and \code{subfunc}.
}
