% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estheta.R
\name{stat_pv}
\alias{stat_pv}
\title{Calculate the imputation mean and variance}
\usage{
stat_pv(PVs_only)
}
\arguments{
\item{PVs_only}{a df given by the result of \code{estheta(est="PVs")}. See\code{\link{estheta}}.}
}
\description{
Calculate the imputation mean and variance
}
